Move the use load module to seperate file
