+++ packages/react-core/src/index.ts
@@ -3,4 +3,5 @@ export * from './scalprum-provider';
 export * from './use-scalprum';
 export * from './scalprum-context';
 export * from './use-module';
+export * from './use-load-module';
 export { ScalprumProvider as default } from './scalprum-provider';
